CEDS-2130 Fix ArrivalSummaryViewSpec
